requirements.txt
# Core data science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Data loading
openpyxl>=3.1.0          # Read Excel files (Food Environment Atlas)
requests>=2.31.0          # NASS API calls
httpx>=0.27.0             # Async API calls (weather, NASS)

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Agentic AI — LangGraph + Archia
langgraph>=0.2.0          # Agent orchestration (stateful graphs)
langchain>=0.3.0
langchain-openai>=0.3.0   # ChatOpenAI for Archia gateway
langchain-community>=0.3.0
openai>=1.12.0            # Archia uses OpenAI-compatible API

# MCP servers (Model Context Protocol)
mcp>=1.0.0                    # FastMCP server framework
langchain-mcp-adapters>=0.2.0 # MCP -> LangChain tool bridge

# Backend API
fastapi>=0.109.0          # REST API for frontend
uvicorn>=0.27.0           # ASGI server

# Utilities
python-dotenv>=1.0.0      # Load .env files
tqdm>=4.65.0              # Progress bars
jupyter>=1.0.0            # Notebooks

# ML prediction & explainability
xgboost>=2.1.0               # Gradient boosting (risk prediction)
shap>=0.46.0                 # SHAP feature importance & explainability
joblib>=1.4.0                # Model serialization

# Web search
duckduckgo-search>=6.0.0     # Real web results (DDGS) for agricultural news/alerts
